backend/requirements.txt
# Essential FastAPI and Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.25.0
python-multipart>=0.0.6

# Database and ORM
supabase>=2.0.0
sqlalchemy>=2.0.0

# Configuration and Environment
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# HTTP Client and Requests
httpx>=0.25.0
requests>=2.31.0

# Google Auth and Earth Engine
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
earthengine-api>=0.1.350

# AI APIs
google-generativeai>=0.3.0
anthropic>=0.18.0
openai>=1.0.0

# Basic utilities
geopy>=2.0.0
python-dateutil>=2.8.0
pytz>=2023.0

# Weather API
pyowm>=3.3.0

# Development and Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Production
gunicorn>=21.0.0
click>=8.0.0

# Rate Limiting
slowapi>=0.1.8

# Error Tracking and Monitoring
sentry-sdk>=1.30.0
psutil>=5.9.0

# Note: Data science packages (numpy, pandas, etc.) can be added later when needed
# They require compilation tools on Windows for Python 3.13 